moved OutputHealth to schema.json
