+++ internal/pkg/dl/output_health.go
@@ -10,32 +10,19 @@ import (
 	"time"
 
 	"github.com/elastic/fleet-server/v7/internal/pkg/bulk"
+	"github.com/elastic/fleet-server/v7/internal/pkg/model"
 	"github.com/gofrs/uuid"
 )
 
-type DataStream struct {
-	Dataset   string `json:"dataset,omitempty"`
-	Type      string `json:"type,omitempty"`
-	Namespace string `json:"namespace,omitempty"`
-}
-
-type OutputHealth struct {
-	Output     string     `json:"output,omitempty"`
-	State      string     `json:"state,omitempty"`
-	Message    string     `json:"message,omitempty"`
-	Timestamp  string     `json:"@timestamp,omitempty"`
-	DataStream DataStream `json:"data_stream,omitempty"`
-}
-
-func CreateOutputHealth(ctx context.Context, bulker bulk.Bulk, doc OutputHealth) error {
+func CreateOutputHealth(ctx context.Context, bulker bulk.Bulk, doc model.OutputHealth) error {
 	return createOutputHealth(ctx, bulker, FleetOutputHealth, doc)
 }
 
-func createOutputHealth(ctx context.Context, bulker bulk.Bulk, index string, doc OutputHealth) error {
+func createOutputHealth(ctx context.Context, bulker bulk.Bulk, index string, doc model.OutputHealth) error {
 	if doc.Timestamp == "" {
 		doc.Timestamp = time.Now().UTC().Format(time.RFC3339)
 	}
-	doc.DataStream = DataStream{
+	doc.DataStream = &model.DataStream{
 		Dataset:   "fleet_server.output_health",
 		Type:      "logs",
 		Namespace: "default",
