+++ internal/pkg/policy/policy_output_test.go
@@ -467,7 +467,7 @@ func TestPolicyRemoteESOutputPrepare(t *testing.T) {
 			Return(&apiKey, nil).Once()
 		bulker.On("CreateAndGetBulker", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(outputBulker, false).Once()
 		bulker.On("Create", mock.Anything, dl.FleetOutputHealth, mock.Anything, mock.MatchedBy(func(body []byte) bool {
-			var doc dl.OutputHealth
+			var doc model.OutputHealth
 			err := json.Unmarshal(body, &doc)
 			if err != nil {
 				t.Fatal(err)
@@ -525,7 +525,7 @@ func TestPolicyRemoteESOutputPrepare(t *testing.T) {
 			Return(apiKey, err).Once()
 		bulker.On("CreateAndGetBulker", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(outputBulker, false).Once()
 		bulker.On("Create", mock.Anything, dl.FleetOutputHealth, mock.Anything, mock.MatchedBy(func(body []byte) bool {
-			var doc dl.OutputHealth
+			var doc model.OutputHealth
 			err := json.Unmarshal(body, &doc)
 			if err != nil {
 				t.Fatal(err)
