+++ model/schema.json
@@ -430,6 +430,43 @@
       ]
     },
 
+    "output_health": {
+      "description": "Output health represents a health state of an output",
+      "type": "object",
+      "properties": {
+        "output": {
+          "type": "string",
+          "description": "Output ID"
+        },
+        "state": {
+          "type": "string",
+          "description": "Health state, can be HEALTHY or DEGRADED"
+        },
+        "message": {
+          "type": "string",
+          "description": "Long state message if unhealthy"
+        },
+        "@timestamp": {
+          "type": "string",
+          "description": "Timestamp of reported state"
+        },
+        "data_stream": {
+          "type": "object",
+          "properties": {
+            "dataset": {
+              "type": "string"
+            },
+            "type": {
+              "type": "string"
+            },
+            "namespace": {
+              "type": "string"
+            }
+          }
+        }
+      }
+    },
+
     "agent": {
       "title": "Agent",
       "description": "An Elastic Agent that has enrolled into Fleet",
